Percent functionality, changed starting money
Added all of the functionality for the winning chances, and made it so
that the user gets 75 chips for winning but each player puts in 25
chips at each turn.

+++ headers/datacontroller.h
@@ -21,7 +21,7 @@ class DataController {
   int turns;
   vector<string> chip;
   int chipRotation; // this is very necessary.
-  vector<int> moneys;
+  vector<long> moneys;
 
 public:
